+++ .gitignore
@@ -6,11 +6,12 @@
 *.json
 *.DS_Store
 CMakeCache.txt
+build
 *.cmake
 .idea
 CMakeFiles
 cmake-build-debug
-colorDetection/cmake-build-debug
+raspi/colorDetection/cmake-build-debug
 CMakeFiles
 kernel_id.h
 kernel_cfg.c
