--- Controller/controller.h
@@ -1,19 +1,0 @@
-//
-// Created by Timo Oeltze on 13.06.22.
-//
-
-#ifndef EES_CONTROLLER_H
-#define EES_CONTROLLER_H
-
-using namespace std;
-//enum Color {red, blue, yellow, green};
-
-class Controller {
-public:
-    int buckets;
-    bool hasConnection;
-    Controller();
-};
-
-
-#endif //EES_CONTROLLER_H
