+++ raspi/colorDetection/tests/test_colordetection.h
@@ -1,0 +1,20 @@
+//
+// Created by Timo Oeltze on 15.06.22.
+//
+
+#ifndef EES_TEST_COLORDETECTION_H
+#define EES_TEST_COLORDETECTION_H
+
+#include <cstdio>
+#include <string>
+#include <iostream>
+#include <sstream>
+
+using namespace std;
+
+class test_colordetection {
+
+};
+
+
+#endif //EES_TEST_COLORDETECTION_H
